Company user policy.

+++ app/Http/Controllers/CompanyUserController.php
@@ -7,25 +7,41 @@ use App\Http\Requests\User\StoreUserRequest;
 use App\Http\Requests\User\UpdateUserRequest;
 use App\Models\Company;
 use App\Models\User;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
 
 class CompanyUserController extends Controller
 {
+    /**
+     * @throws AuthorizationException
+     */
     public function index(Company $company): View
     {
+        $this->authorize('viewAny', $company);
+
         $users = $company->users()->where('role_id', Role::COMPANY_OWNER->value)->get();
 
         return view('companies.users.index', compact('company', 'users'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function create(Company $company): View
     {
+        $this->authorize('create', $company);
+
         return view('companies.users.create', compact('company'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function store(StoreUserRequest $request, Company $company): RedirectResponse
     {
+        $this->authorize('create', $company);
+
         $company->users()->create([
             'name' => $request->input('name'),
             'email' => $request->input('email'),
@@ -36,20 +52,35 @@ class CompanyUserController extends Controller
         return to_route('companies.users.index', $company);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function edit(Company $company, User $user): View
     {
+        $this->authorize('update', $company);
+
         return view('companies.users.edit', compact('company', 'user'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function update(UpdateUserRequest $request, Company $company, User $user): RedirectResponse
     {
+        $this->authorize('update', $company);
+
         $user->update($request->validated());
 
         return to_route('companies.users.index', $company);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(Company $company, User $user): RedirectResponse
     {
+        $this->authorize('delete', $company);
+
         $user->delete();
 
         return to_route('companies.users.index', $company);
